# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    name: DevHelp Team
    url: https://github.com/bloiss/DevHelp
  description: API REST du forum DevHelp — entraide Dev & IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DevHelp API
  version: "1.0"
paths:
...
  /api/v1/conversations:
    get:
      description: Retourne toutes les conversations de l'utilisateur connecté, triées
        par activité récente.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lister les conversations
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Ouvre une conversation existante ou en crée une nouvelle avec un
        autre utilisateur. Retourne la conversation.
      parameters:
      - description: 'user_id: UUID de l''interlocuteur'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ouvrir ou créer une conversation
      tags:
      - messages
  /api/v1/conversations/{id}/accept:
    post:
      description: Accepte une demande de conversation en attente, permettant l'échange
        de messages.
      parameters:
      - description: UUID conversation
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Accepter une demande de conversation
      tags:
      - messages
  /api/v1/conversations/{id}/messages:
    get:
      description: Retourne les messages d'une conversation avec pagination par curseur
        (before timestamp). L'utilisateur doit être participant.
      parameters:
      - description: UUID conversation
        in: path
        name: id
        required: true
        type: string
      - description: Curseur de pagination (RFC3339 timestamp)
        in: query
        name: before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lister les messages d'une conversation
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Envoie un message dans une conversation. Le message peut contenir
        du texte, une pièce jointe, ou un post partagé.
      parameters:
      - description: UUID conversation
        in: path
        name: id
        required: true
        type: string
      - description: Contenu du message (content, attachment_url, attachment_type,
          shared_post_id)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Envoyer un message
      tags:
      - messages
  /api/v1/conversations/{id}/presence:
    get:
      description: Retourne le statut de présence en ligne des autres participants
        de la conversation.
      parameters:
      - description: UUID conversation
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Présence des participants d'une conversation
      tags:
      - messages
  /api/v1/conversations/{id}/read:
    post:
      description: Met à jour le curseur de lecture de l'utilisateur dans la conversation
        pour effacer le badge de messages non lus.
      parameters:
      - description: UUID conversation
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Marquer une conversation comme lue
      tags:
      - messages
  /api/v1/conversations/unread-count:
    get:
      description: Retourne le nombre total de conversations contenant des messages
        non lus pour l'utilisateur connecté.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Nombre de conversations avec messages non lus
      tags:
      - messages
...
securityDefinitions:
  BearerAuth:
    description: Entrez "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"